Move C code over to glslang_util.c
