+++ Makefile.common
@@ -1489,6 +1489,7 @@ ifeq ($(HAVE_SLANG),1)
    OBJ += gfx/drivers_shader/slang_process.o
    OBJ += gfx/drivers_shader/slang_preprocess.o
    OBJ += gfx/drivers_shader/glslang_util.o
+   OBJ += gfx/drivers_shader/glslang_util_cxx.o
    OBJ += gfx/drivers_shader/slang_reflection.o
 endif
